Move location dependency stuff to core module
